Use the string to replace the unsigned long and unsigned long long.

+++ python/lsst/ts/xml/field_info.py
@@ -55,7 +55,7 @@ AVRO_TYPES = {
     "long": "long",
     "long long": "long",
     "unsigned short": "int",
-    "unsigned int": "int",
+    "unsigned int": "long",
     "unsigned long": "long",
     "unsigned long long": "long",
     "float": "float",
